Changed job applications view
